+++ src/ZPublisher/WSGIPublisher.py
@@ -161,7 +161,8 @@ def transaction_pubevents(request, response, tm=transaction.manager):
         try:
             # Raise exception from app if handle-errors is False
             # (set by zope.testbrowser in some cases)
-            if request.environ.get('x-wsgiorg.throw_errors', False):
+            if request.environ.get('x-wsgiorg.throw_errors', False) or \
+               response.debug_mode:
                 reraise(*exc_info)
 
             # Handle exception view